# doc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "M & O bank",
    "description": "Bank api",
    "contact": {
      "email": "maarja.tapsi@khk.ee"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    },
    {
      "url": "http://petstore.swagger.io/v2"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "description": "Creates a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Username already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Username or password is missing"
                }
              }
            }
          },
          "401": {
            "description": "Username already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Username already exist"
                }
              }
            }
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get balance by username",
        "operationId": "getBalanceByUsername",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The balance that needs to be fetched. Use user1 for testing",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBalance"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied",
            "content": {}
          },
          "404": {
            "description": "Balance not found",
            "content": {}
          }
        }
      }
    },
    "/balance/{balanceID}": {
      "put": {
        "tags": [
          "balance"
        ],
        "summary": "Update an existing balance",
        "operationId": "updateBalance",
        "parameters": [
          {
            "name": "balanceID",
            "in": "path",
            "description": "Balance object that needs to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated balance object",
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBalance"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful",
            "content": {}
          },
          "404": {
            "description": "Something went wrong",
            "content": {}
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Creates a user session (log in)",
        "operationId": "login",
        "requestBody": {
          "description": "User credentials used for log in",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Missing username or password"
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid username or password"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Deletes a user session (log out)",
        "parameters": [],
        "responses": {
          "204": {
            "description": "User has been logged out and session has been deleted",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "SessionRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "user1",
          "password": "test"
        }
      },
      "SessionResponse": {
        "required": [
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "required": [
              "_id",
              "username",
              "token",
              "name"
            ],
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "balanceId": {
                "type": "string"
              },
              "token": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "user": {
            "_id": "23",
            "balanceId": "5",
            "username": "user1",
            "token": "9f8e22760d6a96881f9450ce4fb82e27",
            "name": "Tatjana Volkov"
          }
        }
      },
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "User": {
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UpdateBalance": {
        "required": [
          "money"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "string"
          }
        }
      },
      "UserBalance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "name": "Authorization",
        "type": "apiKey",
        "in": "header"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    }
  }
}